working condition
it worked on board of 100*100 and 4*5 threads for 20 iteration, it
worked on all the lif files, if i think about any more optimization or i
will update the project.

we also need to add external documentation and some internal
documentation.

Ex1 file now also create P_output and P_output1 to compare into the the
over output